moved token controller to config, used only for generating tokens
